Replace completed wildcard if the completion is a unique file/dir
If a wildcard completion yields a single (as in, a single file or
directory) possible completion, we now replace the user
input (containing wildcards) with the name of the completed file
or dir. This is a bit inconsistent, as Tab does more than
"completing" in this case, but it's also VERY useful for running
commands that cannot process wildcards.