Support DelayedExpansion for environment variables
PyCmd can now use the /V:ON-OFF flags to enable/disable delayed
variable expansion in the spawned cmd.exe.

The default is ON, as the delayed behavior is much more useful in
practice and also enables proper processing of %ERRORLEVEL%. Normally
it should only be turned OFF for compatibility reasons.