Don't show ERRORLEVEL=141
This errorlevel is returned then a command is piped through a
pager (such as less) and the user quits the pager before all the
output has been consumed. In regular (interactive) use this is NOT an
actual error, and other shells (such as fish) also don't display an
error in the prompt.

Note that the ERRORLEVEL 141 is still present in the environment and
can be displayed w/ echo %ERRORLEVEL%.