Crude mechanism to use bash as a backend (WiP)
On Linux, PyCmd now spawns a bash process and controls its tty so as
to inject commands read from the standard input.

The mechanism is based on pty.spawn, as simply creating the process
with a custom tty won't do it for bash -- it insists on reading from
the controlling terminal.