Linux: propagate dir change from Alt-Up or dir-hist to bash
The directory history changes the CWD of the current (PyCmd) process;
we propagate the change to the underlying bash instance by running a
"cd".

This solution is not great, as it alters the status special
variable ($?) and the special macro !!. A more robust solution would
be to write the CWD into a file and use the PROMPT_COMMAND to apply
it.