Linux: changed default prompt color
Our "relative" scheme for defining colors doesn't work too well in an
ANSI-based console, where the typical approach is to hardcode colors
and tweak the color scheme.

For now, we are switching the prompt color to green and the user@host
part to blue, which resembles the default bash color scheme on Ubuntu.