Linux: detect Esc key pressed vs. special key char sequence
If no subsequent character comes after 0x1B in read_input(), we assume
that this was really the Escape key being pressed (as opposed to a
special character sequence as triggered by arrow keys or Alt-combos).

Note that this is a heuristic and might misbehave in extreme
conditions (e.g. very slow exeucution), and it also makes using the
Esc key to input special sequences harder; these are tradeoffs that
other shells (such as fish) make as well.