Linux: swallow the '\r\n' printed by bash after the command to run
This removes the empty line that used to appear before the command
output.

Note that there is still an empty line printed at the end of the
output -- this is a behavior that was introduced on Windows in order
to match what the original cmd.exe does. It might be a good idea to
remove this as well in order to save screen space (both on Linux and
Windows?)