Moved the pushd_stack out of the InputState
The InputState is for things that influence the behavior of the
editing of the current command. The pushd_stack is used by
run_in_cmd(), and belongs next to dir_hist, tmpfile etc.

This also allows us to unit-test run_in_cmd(); a test has been added
for the newly supported pushd/popd.