Linux: improved performance of output capturing
Capturing the stdout of the spawned bash process now behaves in a
"buffered" way -- bytes are only returned from read_shell when a
newline is found.

An exception is made when Esc is detected -- in this case we disable
the buffering so that we immediately flush any potential escape
sequences; buffering is re-enabled at the next newline.

This is a significant improvement, but still orders of magnitude
slower than "native" bash -- as all the output goes through read_shell
and is ultimately processed in Python code! In practice this is not so
bad, as PyCmd is geared towards interactive use.