Linux: improved performance of output capturing
Capturing the stdout of the spawned bash process line now happens
periodically in a "framerate-based" approach.

This is a significant improvement, but still orders of magnitude
slower than "native" bash -- as all the output goes through read_shell
and is ultimately processed in Python code! In practice this is not so
bad, as PyCmd is geared towards interactive use.