Linux: handle Ctrl-Enter in keymap translation
In a standard console, Ctrl-Enter is read as 0x0A (as opposed to
simple Enter == 0x0D).

But when running under X11 (i.e. in a graphical terminal emulator),
Ctrl-Enter registers to the tty as a simple Enter. We resort to using
xlib in order to detect that a Ctrl modifier is pressed (and also
account for the ~common case when CapsLock is remapped to Ctrl!).