Use os.scandir instead of os.listdir for improved performance
For improved performance, we switch to using os.scandir when
completing -- scandir caches the "is_file" and "is_dir" attributes and
thus saves a lot of system calls later.

This drastically improves the responsiveness of Tab-completion AND
auto-suggestion for directories containing many files and/or
subdirectories.

See https://peps.python.org/pep-0471/ for a detailed explanation.