Disable line clearing (Crl-U, Ctrl-K) for the first command
In order to "erase" keys typed while a bash runs a command that does
not consume stdin (and therefore become bash input) we send a clearing
sequence (Ctrl-U, Ctrl-K) before sending bash the command to execute.

For unknown reasons, doing this immediately after starting bash (i.e.
before the first command) sometimes makes bash (non-deterministically!)
issue weird prompt repainting output and this messes up our prompt
detection heuristic.

For lack of a true fix (or even understading) for this issue, we simply
disable the clearing sequence when sending the first command to bash
(clearing is not needed in this case anyway, as there can be no overtype
at this time)